requirements.txt
firecrawl
pandas
python-dotenv
tenacity
streamlit